requirements.txt
fastapi
pydantic
pandas
numpy
python-docx
xlrd
openpyxl